KidController + test disable foreign key from tables

+++ app/Http/Controllers/KidController.php
@@ -7,6 +7,7 @@ use Illuminate\Support\Facades\Auth;
 use Validator;
 use App\Kid;
 use App\User;
+use App\Category;
 
 class KidController extends Controller
 {
@@ -14,6 +15,10 @@ class KidController extends Controller
     public function index()
     {
         $k = Kid::all();
+
+        if($k->user_id != Auth::id()){
+            abort(404);
+        }
         
          return response()->json([
              "message" => "success",
@@ -29,10 +34,12 @@ class KidController extends Controller
         //         'error' => 'ID non renseignée ou invalide'
         //     ]);      
         // }
-    
 
         // $k = Kid::find($req->id);
         $k = Kid::findOrFail($id);
+        if($k->user_id != Auth::id()){
+            abort(404);
+        }
 
         return response()->json([
             'success' => true,
@@ -71,15 +78,12 @@ class KidController extends Controller
         $k->user_id = Auth::id();
 
         $k->save();
+        // Passage par une table pivot
+        $k->categories()->attach($req->categories);
 
         return response()->json([
-            'success' => true,
-            'id' => $k->id,
-            'name' => $k->name,
-            'date_of_birth' => $k->date_of_birth,
-            'avatar' => $k->avatar,
-            'categories' => $k->categories
-        ], 201);
+            "data" => $k
+        ], 200);
     }
 
     public function update(Request $req, $id)
@@ -98,6 +102,7 @@ class KidController extends Controller
         $k->categories = implode(',', $req->categories);
 
         $k->save();
+        $k->categories()->sync($request->category_id, false);
 
         return response()->json($k, 200);
     }
@@ -112,6 +117,10 @@ class KidController extends Controller
         // }
 
         $k = Kid::findOrFail($id);
+        if($k->user_id != Auth::id()){
+            abort(404);
+        }
+            
         // $k = Kid::find($req->id);
         $k->delete();
